Added option to dump to app-root/data/.balloonstats.log
OpenShift does not normally allow app code to write to the home dir
which is where the .balloonstats.log file is normally located (the
working directory of an OpenShift app defaults to $HOME). If value
approot is supplied as an option to the -agentpath argument then the
log file is relocated to $CWD/app-root/data, a directory which is
writeable by app code e.g.

  java -agentpath:/path/to/libballoon.so=verbose,approot

n.b. option approot is incompatible with the sysout option which
redirects log output to System.out. if both options are supplied the
last one passed will take effect.